// src/Proto/v1alpha.ts
export namespace grpc.reflection.v1alpha {
// ...
    interface GrpcChunkServiceInterface<P, R> {
        (metadata: grpc_1.Metadata, options?: grpc_1.CallOptions): grpc_1.ClientDuplexStream<P, R>;
        (options?: grpc_1.CallOptions): grpc_1.ClientDuplexStream<P, R>;
    }
    interface GrpcPromiseServiceInterface<P, R> {
        (message: P, metadata: grpc_1.Metadata, options?: grpc_1.CallOptions): Promise<R>;
        (message: P, options?: grpc_1.CallOptions): Promise<R>;
    }
    export abstract class UnimplementedServerReflectionService {
        static definition = {
            ServerReflectionInfo: {
                path: "/grpc.reflection.v1alpha.ServerReflection/ServerReflectionInfo",
                requestStream: true,
                responseStream: true,
                requestSerialize: (message: ServerReflectionRequest) => Buffer.from(message.serialize()),
                requestDeserialize: (bytes: Buffer) => ServerReflectionRequest.deserialize(new Uint8Array(bytes)),
                responseSerialize: (message: ServerReflectionResponse) => Buffer.from(message.serialize()),
                responseDeserialize: (bytes: Buffer) => ServerReflectionResponse.deserialize(new Uint8Array(bytes))
            }
        };
        [method: string]: grpc_1.UntypedHandleCall;
        abstract ServerReflectionInfo(call: grpc_1.ServerDuplexStream<ServerReflectionRequest, ServerReflectionResponse>): void;
    }
    export class ServerReflectionClient extends grpc_1.makeGenericClientConstructor(UnimplementedServerReflectionService.definition, "ServerReflection", {}) {
        constructor(address: string, credentials: grpc_1.ChannelCredentials, options?: Partial<grpc_1.ChannelOptions>) {
            super(address, credentials, options);
        }
        ServerReflectionInfo: GrpcChunkServiceInterface<ServerReflectionRequest, ServerReflectionResponse> = (metadata?: grpc_1.Metadata | grpc_1.CallOptions, options?: grpc_1.CallOptions): grpc_1.ClientDuplexStream<ServerReflectionRequest, ServerReflectionResponse> => {
            return super.ServerReflectionInfo(metadata, options);
        };
    }
}